11382: 꼬마 정민

+++ JAVA/백준/01. 입출력과 사칙연산/B11382_꼬마정민.java
@@ -9,9 +9,9 @@ public class B11382_꼬마정민 {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine(), " ");
 
-        int A = Integer.parseInt(st.nextToken());
-        int B = Integer.parseInt(st.nextToken());
-        int C = Integer.parseInt(st.nextToken());
+        long A = Long.parseLong(st.nextToken());
+        long B = Long.parseLong(st.nextToken());
+        long C = Long.parseLong(st.nextToken());
 
         System.out.println(A+B+C);
     }
